services/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Simple FastAPI server with health endpoint"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "."}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.0"
uvicorn = "^0.28.0"
pydantic = "^2.10.0"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.0"}

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^7.4.0"
black = "^24.2.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
